[UPDATE 0.24.1] Restructuring the project
