+++ .gitignore
@@ -1,6 +1,7 @@
 /.vscode/
 /__pycache__/
 /Routes/__pycache__/
+/Models/__pycache__/
 /venv/
 /Environment.py
 /Cache/
